+++ src/routes/config.js
@@ -3542,7 +3542,7 @@ const Configuration = () => {
                 <Grid container spacing={2} rowSpacing={-0.5}>
                     <Grid item xs={12} md={6}>
                         <Typography variant="body2">
-                            Client: 3.4.5 (build.{buildhash})
+                            Client: 3.4.6 (build.{buildhash})
                         </Typography>
                     </Grid>
                     <Grid item xs={12} md={6}>